pyproject.toml
[project]
name = "i2e-project"
version = "0.1.0"
description = "NASA Systems Engineering Handbook QA System"
requires-python = ">=3.11"
dependencies = [
    "pymupdf>=1.24.0",         # PDF text + image extraction
    "pdfplumber>=0.11.0",      # Table extraction
    "openai>=1.30.0",          # text-embedding-3-small
    "chromadb>=0.5.0",         # local vector store
    "rank-bm25>=0.2.2",        # BM25 keyword index
    "anthropic>=0.28.0",       # Claude API
    "streamlit>=1.35.0",       # chat UI
    "python-dotenv>=1.0.0",    # .env for API keys
    "tqdm>=4.66.0",            # progress bars
    "pandas>=2.3.3",
]

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
